src/prepare/prepare.py
import io
import os
import random
import re
import sys
import xml.etree.ElementTree

import yaml

from prepare_data import *

# taken from https://github.com/optas/latent_3d_points/blob/8e8f29f8124ed5fc59439e8551ba7ef7567c9a37/src/in_out.py
synsetid_to_cate = {
    "02691156": "airplane",
    "02773838": "bag",
    "02801938": "basket",
    "02808440": "bathtub",
    "02818832": "bed",
    "02828884": "bench",
    "02876657": "bottle",
    "02880940": "bowl",
    "02924116": "bus",
    "02933112": "cabinet",
    "02747177": "can",
    "02942699": "camera",
    "02954340": "cap",
    "02958343": "car",
    "03001627": "chair",
    "03046257": "clock",
    "03207941": "dishwasher",
    "03211117": "monitor",
    "04379243": "table",
    "04401088": "telephone",
    "02946921": "tin_can",
    "04460130": "tower",
    "04468005": "train",
    "03085013": "keyboard",
    "03261776": "earphone",
    "03325088": "faucet",
    "03337140": "file",
    "03467517": "guitar",
    "03513137": "helmet",
    "03593526": "jar",
    "03624134": "knife",
    "03636649": "lamp",
    "03642806": "laptop",
    "03691459": "speaker",
    "03710193": "mailbox",
    "03759954": "microphone",
    "03761084": "microwave",
    "03790512": "motorcycle",
    "03797390": "mug",
    "03928116": "piano",
    "03938244": "pillow",
    "03948459": "pistol",
    "03991062": "pot",
    "04004475": "printer",
    "04074963": "remote_control",
    "04090263": "rifle",
    "04099429": "rocket",
    "04225987": "skateboard",
    "04256520": "sofa",
    "04330267": "stove",
    "04530566": "vessel",
    "04554684": "washer",
    "02992529": "cellphone",
    "02843684": "birdhouse",
    "02871439": "bookshelf",
    # '02858304': 'boat', no boat in our dataset, merged into vessels
    # '02834778': 'bicycle', not in our taxonomy
}
cate_to_synsetid = {v: k for k, v in synsetid_to_cate.items()}

# reads in parameters from params.yaml
params = yaml.safe_load(open(os.path.join(HOME_PATH, "params.yaml")))["prepare"]

DATASET = params["dataset"]
CATEGORY = (
    cate_to_synsetid[params["category"]]
    if DATASET == "shapenet"
    else params["category"]
)
NUM_POINT = params["num_point"]
VAR_RANGE = params["var_range"]

save_folder = os.path.join(HOME_PATH, "data/clean", DATASET, params["category"])
data_folder = os.path.join(
    HOME_PATH, "data/raw", DATASET, CATEGORY, "points/" if DATASET == "shapenet" else ""
)

os.makedirs(save_folder, exist_ok=True)

preprocess_and_save(folder=data_folder, save_path=save_folder)
